fixed most of the iceflow stuff, implemented a search algorithm which spirals out from a starting point until it finds a x-y pair with the iceflow vector which is pretty cool.
did some work on the crossover notebook and implemented a method of dealing with bad crossover points (and fake ones) by checking to see if the difference in headings on the legs is more than a set angle.

Something in the plot code seems to be making one of the segments not plot in the right place on the map but I'll get to that later. I think the other plot is fine - no it is not!