Does anyone read this? I'll be at the coffee shop accross the street.
Started working on a function to append two flights to each other. Also started on a function to just loop through a whole season directory and do that but it looks like the 2022 Basler flight got saved in an old version of matlab file so it won't work with h5py or the lovely h5py layerize function I've been using.

Also started toying with fitting functions for the angle plots. So far the blue polynomial seems best.